Add files via upload
project not complete but 1 or 2 more days would have done it. credential details are hardcoded. the token expires so i would have to enter a new one for the other functions to work. i know how to deserialize json to work with the data easier but had no time! sorry.  i lost time because i saw the email late after the online exam. my fault. just wish i had the time to complete this interesting project.