Add files via upload
Hello
sadly this may be the final commit! A beta version of the App. I hope to be called for the interview where i can be asked about it...
anyway, in the beta version i tested all functions and they work . i did not have time to use try constructs in some cases. and  validations.

all the best.
i will appreciate the opportunity to work for MRA,